requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
python-dotenv==1.2.1
PyJWT==2.10.1
sqlparse==0.5.3
requests==2.32.5
yt-dlp==2025.10.22
torch==2.2.2
numpy==1.26.4
openai-whisper==20250625
google-genai==1.56.0
tqdm==4.67.1
